per-thread op recording, bounds checks in TriMesh
Made operation logger use a RAII recording object. To allow for
recording between multiple functions TriMesh holds a a per-thread
weak_ptr to a recording object. Operations create a recorder shared_ptr
and destroy it when they end, which both constructs a payload and writes
the payload to the logger respectively. All recording due to internal
function calls are added to the payload by checking for the weak_ptr.

If an operation fails then the op needs to explicitly declare that it
needs to be cancelled. All access to recorders is hidden behind a
p_opeartion_logger to avoid leaking access to the
tbb::enumerable_thread_specific to circumstances where recording is not
necessary.

This structure is desirable because topology and attribute updates can
happen outside of a single op's invocation and the recorder object needs
to be made available to those objects. This solution is a bit more
intricate than doing parameter passing, but I thought it'd be better to
keep function interfaces lighter by bundling the logging/recording in
TriMesh.